Add "Delete confirm" dialog

+++ resources/views/pluginfunctions/index.blade.php
@@ -39,7 +39,7 @@
                         <td>
                             <div class="btn-group">
                                 <a href="/dashboard/pluginfunctions/{{ $pfunction->id }}/edit" class="btn btn-secondary btn-sm" role="button">edit</a>
-                                {!! Form::open(['route' => ['pluginfunctions.destroy', $pfunction->id], 'method'  => 'DELETE']) !!}
+                                {!! Form::open(['route' => ['pluginfunctions.destroy', $pfunction->id], 'method'  => 'DELETE', 'onclick' => 'return confirm("Delete '.$pfunction->name.' ?")' ]) !!}
                                 {!! Form::submit('delete', ['class' => 'btn btn-danger btn-sm']) !!}
                                 {!! Form::close() !!}
                             </div>
